requirements.txt
psycopg2-binary==2.9.9
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
packaging==23.1
PyJWT==2.8.0
pyotp==2.9.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
django-markdown==0.8.4
django-filter==24.3
django-rest-framework-social-oauth2==1.2.0
django-oauth-toolkit==2.4.0
